Dedupe videos. Add VideoFeedViewModel for managing video feed operations
Introduced `VideoFeedViewModel` to handle video fetching, caching, and playback queue management. Extracted related logic from `VideoFeedView.swift` to the new class, improving modularity and maintainability. Adjusted project files and removed redundant code to integrate the new view model seamlessly.
