+++ Arkavo/Arkavo/VideoFeedView.swift
@@ -4,34 +4,6 @@ import Foundation
 import SwiftData
 import SwiftUI
 
-// MARK: - Models
-
-struct Video: Identifiable {
-    let id: String
-    let url: URL
-    let contributors: [Contributor]
-    let description: String
-
-    static func from(uploadResult: UploadResult, contributors: [Contributor]) -> Video {
-        Video(
-            id: uploadResult.id,
-            url: URL(string: uploadResult.playbackURL)!,
-            contributors: contributors,
-            description: Date().ISO8601Format()
-        )
-    }
-
-    static func from(thought: Thought) -> Video? {
-        let url = URL(string: "nano://\(thought.publicID.base58EncodedString)/")!
-        return Video(
-            id: thought.id.uuidString,
-            url: url,
-            contributors: thought.metadata.contributors,
-            description: thought.metadata.createdAt.ISO8601Format()
-        )
-    }
-}
-
 // MARK: - Main View
 
 struct VideoContentView: View {
@@ -39,7 +11,6 @@ struct VideoContentView: View {
     @StateObject private var feedViewModel: VideoFeedViewModel
 
     init() {
-        // Initialize with an empty view model, it will be properly configured in onAppear
         _feedViewModel = StateObject(wrappedValue: ViewModelFactory.shared.makeVideoFeedViewModel())
     }
 
@@ -202,7 +173,6 @@ struct ContributorsView: View {
 struct VideoPlayerView: View {
     @EnvironmentObject var sharedState: SharedState
     @ObservedObject var viewModel: VideoFeedViewModel
-    @State private var dragOffset = CGSize.zero
     let video: Video
     let size: CGSize
     private let swipeThreshold: CGFloat = 50
@@ -278,23 +248,17 @@ struct VideoPlayerView: View {
             }
             .gesture(
                 DragGesture()
-                    .onChanged { gesture in
-                        dragOffset = gesture.translation
-                    }
                     .onEnded { gesture in
                         let verticalMovement = gesture.translation.height
                         if abs(verticalMovement) > swipeThreshold {
-                            if verticalMovement > 0, viewModel.currentVideoIndex > 0 {
-                                withAnimation {
-                                    viewModel.currentVideoIndex -= 1
-                                }
-                            } else if verticalMovement < 0, viewModel.currentVideoIndex < viewModel.videos.count - 1 {
-                                withAnimation {
-                                    viewModel.currentVideoIndex += 1
+                            Task {
+                                if verticalMovement < 0 {
+                                    await viewModel.handleSwipe(.up)
+                                } else {
+                                    await viewModel.handleSwipe(.down)
                                 }
                             }
                         }
-                        dragOffset = .zero
                     }
             )
         }
@@ -471,297 +435,4 @@ struct PlayerContainerView: UIViewRepresentable {
 enum VideoStreamError: Error {
     case noVideoStream
     case invalidStream
-}
-
-@MainActor
-final class VideoFeedViewModel: ObservableObject, VideoFeedUpdating {
-    let client: ArkavoClient
-    let account: Account
-    let profile: Profile
-    @Published var videos: [Video] = []
-    @Published var currentVideoIndex = 0
-    @Published var isLoading = false
-    @Published var error: Error?
-    @Published var connectionState: ArkavoClientState = .disconnected
-    let playerManager = VideoPlayerManager()
-    private var notificationObservers: [NSObjectProtocol] = []
-
-    init(client: ArkavoClient, account: Account, profile: Profile) {
-        self.client = client
-        self.account = account
-        self.profile = profile
-        setupNotifications()
-        // Load initial videos
-        Task {
-            await loadVideos()
-        }
-    }
-
-    private func setupNotifications() {
-        print("VideoFeedViewModel: setupNotifications")
-        // Clean up any existing observers
-        notificationObservers.forEach { NotificationCenter.default.removeObserver($0) }
-        notificationObservers.removeAll()
-
-        // Connection state changes
-        let stateObserver = NotificationCenter.default.addObserver(
-            forName: .arkavoClientStateChanged,
-            object: nil,
-            queue: nil
-        ) { [weak self] notification in
-            guard let state = notification.userInfo?["state"] as? ArkavoClientState else { return }
-            Task { @MainActor [weak self] in
-                self?.connectionState = state
-            }
-        }
-        notificationObservers.append(stateObserver)
-
-        // Decrypted message handling
-        let messageObserver = NotificationCenter.default.addObserver(
-            forName: .messageDecrypted,
-            object: nil,
-            queue: nil
-        ) { [weak self] notification in
-            guard let data = notification.userInfo?["data"] as? Data,
-                  let policy = notification.userInfo?["policy"] as? ArkavoPolicy else { return }
-
-            Task { @MainActor [weak self] in
-                await self?.handleDecryptedMessage(data: data, policy: policy)
-            }
-        }
-        notificationObservers.append(messageObserver)
-
-        // Error handling
-        let errorObserver = NotificationCenter.default.addObserver(
-            forName: .messageHandlingError,
-            object: nil,
-            queue: nil
-        ) { [weak self] notification in
-            guard let error = notification.userInfo?["error"] as? Error else { return }
-            Task { @MainActor [weak self] in
-                self?.error = error
-            }
-        }
-        notificationObservers.append(errorObserver)
-    }
-
-    private func loadVideos() async {
-        isLoading = true
-        defer { isLoading = false }
-
-        // First try to load from stream
-        if let videoStream = try? await getOrCreateVideoStream() {
-            // Load any cached messages first
-            let queueManager = ViewModelFactory.shared.serviceLocator.resolve() as MessageQueueManager
-            let router = ViewModelFactory.shared.serviceLocator.resolve() as ArkavoMessageRouter
-            if let (messageId, message) = queueManager.getNextMessage(
-                ofType: 0x05,
-                forStream: videoStream.publicID
-            ) {
-                do {
-                    try await router.processMessage(message.data, messageId: messageId)
-                } catch {
-                    print("Failed to process cached message: \(error)")
-                }
-            }
-
-            // If still no videos, load from account video-stream thoughts
-            if videos.isEmpty {
-                if let thought = videoStream.thoughts.last(where: { $0.metadata.mediaType == .video }) {
-                    try? await router.processMessage(thought.nano, messageId: thought.id)
-                }
-            }
-            // Wait for a limited time if we still have no videos
-            let maxAttempts = 10 // 1 second total wait
-            var attempts = 0
-            while videos.isEmpty, attempts < maxAttempts {
-                try? await Task.sleep(nanoseconds: 100_000_000)
-                attempts += 1
-            }
-        }
-    }
-
-    deinit {
-        // Clean up observers
-        notificationObservers.forEach { NotificationCenter.default.removeObserver($0) }
-    }
-
-    private func handleDecryptedMessage(data: Data, policy: ArkavoPolicy) async {
-        do {
-            print("\nHandling decrypted video message:")
-            print("- Data size: \(data.count)")
-            print("- Policy type: \(policy.type)")
-
-            // Verify this is a video message based on policy
-            guard policy.type == .videoFrame else {
-                print("❌ Incorrect policy type")
-                return
-            }
-
-            // Create a temporary file URL in the cache directory for the video data
-            let fileManager = FileManager.default
-            let cacheDir = fileManager.urls(for: .cachesDirectory, in: .userDomainMask)[0]
-            let videoFileName = UUID().uuidString + ".mp4" // Or appropriate extension
-            let videoFileURL = cacheDir.appendingPathComponent(videoFileName)
-
-            // Write the video data to the cache file
-            try data.write(to: videoFileURL)
-            print("✅ Wrote video data to cache: \(videoFileURL)")
-
-            // Analyze the video file after writing
-            let asset = AVURLAsset(url: videoFileURL)
-            if let videoTrack = try await asset.loadTracks(withMediaType: .video).first {
-                let naturalSize = try await videoTrack.load(.naturalSize)
-                let transform = try await videoTrack.load(.preferredTransform)
-                let videoAngle = atan2(transform.b, transform.a)
-
-                print("\n📼 Decrypted Video Analysis:")
-                print("- File size: \(data.count) bytes")
-                print("- Natural size: \(naturalSize)")
-                print("- Aspect ratio: \(naturalSize.width / naturalSize.height)")
-                print("- Transform angle: \(videoAngle * 180 / .pi)°")
-                print("- Transform matrix: \(transform)")
-            }
-
-            // Create new video object using the cached file URL
-            // FIXME: use ArkavoPolicy.Metadata
-            let video = Video(
-                id: UUID().uuidString,
-                url: videoFileURL,
-                contributors: [], // Add contributors if available in policy
-                description: policy.type.rawValue
-            )
-
-            // Update UI
-            await MainActor.run {
-                print("Adding video to feed")
-                videos.insert(video, at: 0)
-                if videos.count == 1 {
-                    print("Preloading first video")
-                    preloadVideo(url: video.url)
-                }
-            }
-
-        } catch {
-            print("❌ Error processing video: \(error)")
-            await MainActor.run {
-                self.error = error
-            }
-        }
-    }
-
-    func streams() -> [Stream] {
-        let streams = account.streams.dropFirst(2)
-        return Array(streams)
-    }
-
-    func cleanupOldCacheFiles() {
-        let fileManager = FileManager.default
-        guard let cacheDir = fileManager.urls(for: .cachesDirectory, in: .userDomainMask).first else { return }
-
-        do {
-            let cacheContents = try fileManager.contentsOfDirectory(
-                at: cacheDir,
-                includingPropertiesForKeys: [.creationDateKey],
-                options: [.skipsHiddenFiles]
-            )
-
-            // Keep only the 20 most recent videos
-            let oldFiles = cacheContents
-                .filter { $0.pathExtension == "mp4" }
-                .sorted { url1, url2 -> Bool in
-                    let date1 = try? url1.resourceValues(forKeys: [.creationDateKey]).creationDate ?? Date.distantPast
-                    let date2 = try? url2.resourceValues(forKeys: [.creationDateKey]).creationDate ?? Date.distantPast
-                    return date1! > date2!
-                }
-                .dropFirst(20)
-
-            for fileURL in oldFiles {
-                try? fileManager.removeItem(at: fileURL)
-            }
-        } catch {
-            print("Error cleaning cache: \(error)")
-        }
-    }
-
-    func addNewVideo(from uploadResult: UploadResult, contributors: [Contributor]) {
-        Task {
-            do {
-                // Get or create video stream
-                let videoStream = try await getOrCreateVideoStream()
-
-                // Create new video
-                let newVideo = Video.from(uploadResult: uploadResult, contributors: contributors)
-
-                // Create thought for the video
-                let metadata = Thought.Metadata(
-                    creatorPublicID: profile.publicID,
-                    streamPublicID: videoStream.publicID,
-                    mediaType: .video,
-                    createdAt: Date(),
-                    contributors: contributors
-                )
-
-                // Convert video URL to data for storage
-                let videoData = newVideo.url.absoluteString.data(using: .utf8) ?? Data()
-                let thought = Thought(nano: videoData, metadata: metadata)
-
-                // Add thought to stream
-                videoStream.thoughts.append(thought)
-
-                // Save changes to persistence
-                try await PersistenceController.shared.saveChanges()
-
-                // Update UI
-                await MainActor.run {
-                    videos.insert(newVideo, at: 0)
-                    currentVideoIndex = 0
-                }
-
-                // Preload video
-                try? await playerManager.preloadVideo(url: newVideo.url)
-
-            } catch VideoStreamError.noVideoStream {
-                print("No video stream exists and couldn't create one")
-                self.error = VideoStreamError.noVideoStream
-            } catch {
-                print("Error adding video: \(error)")
-                self.error = error
-            }
-        }
-    }
-
-    func getOrCreateVideoStream() async throws -> Stream {
-        // First check for existing video stream
-        if let existingStream = account.streams.first(where: { stream in
-            stream.source?.metadata.mediaType == .video
-        }) {
-            return existingStream
-        }
-        throw VideoStreamError.noVideoStream
-    }
-
-    func preloadVideo(url: URL) {
-        Task {
-            do {
-                try await playerManager.preloadVideo(url: url)
-            } catch {
-                self.error = error
-            }
-        }
-    }
-}
-
-@MainActor
-protocol VideoFeedUpdating: AnyObject {
-    func addNewVideo(from result: UploadResult, contributors: [Contributor])
-    func preloadVideo(url: URL)
-}
-
-// Extension to provide default implementation
-@MainActor
-extension VideoFeedUpdating {
-    func preloadVideo(url _: URL) {
-        // Optional default implementation
-    }
 }
